Create a k8s service pointing to a particular deployment
Create an additional service per app (gateway_service). For canary deployments, gateway service will point to the newest version once it has weight. In the event of a rollback, the weight of the most recent version will go to 0 and gateway service will go back to the previous version.

Test by deploying the controller from this branch and creating an app. Verify that the gateway service has been created by running `kubectl get services -A` and check for a service with a name app-<app-name>