Scale target deployment for canary (#151)
Scale process' units based on the current weight of deployment and specified target values per process. The Target map has been added to the CanarySpec, which is used to specify the target number of units per process. During the canary deployment, the two deployments' processes should add up to the target value (per process). The final deployment should have the same value as the target (per process).

In the previous deployment, if a process is not present in the target, the number of units will not be affected during the canary deployment, but will end up being removed (along with the rest of the deployment) if the canary deployment successfully completes.

In the updated deployment, if a process is not present in the target list, the process will be given a default value of 1.