[PLAY-648] Multi platform images (#289)
Idea is to build multiplatform image, but since cross platform building ARM64 image is really slow (20mins) we will also build old way.

jobs:

build-single-platform - it's old way of building images (no qemu/buildx)
this job runs on PR requests, dev* branch, and also master branch
it is fast, and it will be first to built and pushed to docker hub, so that image that works for most situations is available quickly.

trivy security scans, wait for build-single-platform to complete, and run security scan based on that. It also triggers on schedule, and uses that image

build-multiplatform - this is new really slow way of building image
this job runs only on master
it is slow, and it will overwrite build-single-platform image once it completes. It should work for all distributions
merge waits for build-multiplatform job and it combines all platforms into single tag